backend/requirements.txt
Django>=3.2,<4.0
djangorestframework<3.14.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
gunicorn>=20.1.0
whitenoise>=5.3.0
django-cors-headers<4.0